pyproject.toml
[tool.poetry]
name = "qortex"
version = "0.1.0"
description = ""
authors = ["subbotinav <subbotinav"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
drf-yasg = "^1.21.4"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
